NEW added better support for CPBrowser
The previous test for focus when drawing selection color was not
valid for CPBrowser as it uses a private subclass of tableview and
does not accept firstResponder.

To accomodate this a new private _isFocused() method has been added to tableview
and modified for the private subclass.

CPBrowser was also updted to force redraw of its previous/current tableview
when lastColumn changes.

CPBrowser behaves like cocoa, where only the currently selected column shows
the blue highlight and all others are grey.